move bundle fetcher to node/bundle instead of chain/actors
